Bugfix/andkr 12 qa v183 0412 02

+++ dailyHOTEL/app/src/main/java/com/twoheart/dailyhotel/screen/booking/list/BookingListAdapter.java
@@ -126,7 +126,7 @@ public class BookingListAdapter extends ArrayAdapter<Booking> implements PinnedS
         {
             Integer tag = (Integer) convertView.getTag();
 
-            if (tag == null || tag.intValue() != booking.type)
+            if (tag == null || tag != booking.type)
             {
                 convertView = null;
             }
